backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
networkx
openai